unify FlinkBatchCost and FlinkStreamCost to FlinkCost
