+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/cost/FlinkCostFactory.scala
@@ -18,18 +18,32 @@
 
 package org.apache.flink.table.plan.cost
 
-import org.apache.calcite.plan.{RelOptCost, RelOptCostFactory}
+import org.apache.calcite.plan.RelOptCost
 
 /**
-  * A [[RelOptCostFactory]] that adds makeCost methods with network and memory parameters.
+  * This class is based on Apache Calcite's [[org.apache.calcite.plan.volcano.VolcanoCost#Factory]].
   */
-trait FlinkCostFactory extends RelOptCostFactory {
+class FlinkCostFactory extends FlinkCostFactoryBase {
 
-  def makeCost(
-    rowCount: Double,
-    cpu: Double,
-    io: Double,
-    network: Double,
-    memory: Double): RelOptCost
+  override def makeCost(
+      rowCount: Double,
+      cpu: Double,
+      io: Double,
+      network: Double,
+      memory: Double): RelOptCost = {
+    new FlinkCost(rowCount, cpu, io, network, memory)
+  }
+
+  override def makeCost(dRows: Double, dCpu: Double, dIo: Double): RelOptCost = {
+    new FlinkCost(dRows, dCpu, dIo, 0.0, 0.0)
+  }
+
+  override def makeHugeCost: RelOptCost = FlinkCost.Huge
+
+  override def makeInfiniteCost: RelOptCost = FlinkCost.Infinity
+
+  override def makeTinyCost: RelOptCost = FlinkCost.Tiny
+
+  override def makeZeroCost: RelOptCost = FlinkCost.Zero
 
 }
