+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/cost/FlinkCostFactoryBase.scala
@@ -18,32 +18,18 @@
 
 package org.apache.flink.table.plan.cost
 
-import org.apache.calcite.plan.RelOptCost
+import org.apache.calcite.plan.{RelOptCost, RelOptCostFactory}
 
 /**
-  * This class is based on Apache Calcite's [[org.apache.calcite.plan.volcano.VolcanoCost#Factory]].
+  * A [[RelOptCostFactory]] that adds makeCost methods with network and memory parameters.
   */
-class FlinkBatchCostFactory extends FlinkCostFactory {
+trait FlinkCostFactoryBase extends RelOptCostFactory {
 
-  override def makeCost(
-      rowCount: Double,
-      cpu: Double,
-      io: Double,
-      network: Double,
-      memory: Double): RelOptCost = {
-    new FlinkBatchCost(rowCount, cpu, io, network, memory)
-  }
-
-  override def makeCost(dRows: Double, dCpu: Double, dIo: Double): RelOptCost = {
-    new FlinkBatchCost(dRows, dCpu, dIo, 0.0, 0.0)
-  }
-
-  override def makeHugeCost: RelOptCost = FlinkBatchCost.Huge
-
-  override def makeInfiniteCost: RelOptCost = FlinkBatchCost.Infinity
-
-  override def makeTinyCost: RelOptCost = FlinkBatchCost.Tiny
-
-  override def makeZeroCost: RelOptCost = FlinkBatchCost.Zero
+  def makeCost(
+    rowCount: Double,
+    cpu: Double,
+    io: Double,
+    network: Double,
+    memory: Double): RelOptCost
 
 }
